Automatic program generation
It is now possilbe to take a .cl code and add it into the build process.
The build process will generate a C++ class (in a .cpp and .h files).
This class will be compiled and can be used directly in the application
code. The validity of the .cl file is checked via oclc.